further development for loading and using shared libraries cleanup of header files  - created C/C++ headers
git-svn-id: https://subversion.assembla.com/svn/zappralot/trunk/libfortres@593 c37337d0-301c-4318-a67e-7f75229f9ee7

+++ include/fortres/StringRef.hpp
@@ -3,7 +3,7 @@
 
 #include <string>
 #include <string.h>
-#include "fortres/portdef.h"
+#include <stdint.h>
 
 #pragma pack(push, 4)
 class StringRef
@@ -48,8 +48,8 @@ class StringRef
       }
 
   private:
-    char     *_ref;
-    uint32_t  _len;
+    char   *_ref;
+    size_t  _len;
 };
 #pragma pack(pop)
